Move ExtensionsHandler from crac-impl to crac-api
