+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/ExtensionsHandler.java
@@ -1,13 +1,13 @@
 /*
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ *  Copyright (c) 2020, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *  License, v. 2.0. If a copy of the MPL was not distributed with this
+ *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
  */
 
-package com.farao_community.farao.data.crac_impl.json;
+package com.farao_community.farao.data.crac_api;
 
-import com.farao_community.farao.data.crac_api.Identifiable;
 import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 import com.powsybl.commons.extensions.Extension;
@@ -27,7 +27,7 @@ public final class ExtensionsHandler {
             Suppliers.memoize(() -> ExtensionProviders.createProvider(ExtensionSerializer.class));
 
     /**
-     * Gets the known Cnec extension serializers.
+     * Gets the known extension serializers.
      */
     public static ExtensionProviders<ExtensionSerializer> getExtensionsSerializers() {
         return SERIALIZER_SUPPLIER.get();
