+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/json/deserializers/SimpleCracDeserializer.java
@@ -10,7 +10,7 @@ package com.farao_community.farao.data.crac_impl.json.deserializers;
 import com.farao_community.farao.commons.FaraoException;
 import com.farao_community.farao.data.crac_api.*;
 import com.farao_community.farao.data.crac_impl.SimpleCrac;
-import com.farao_community.farao.data.crac_impl.json.ExtensionsHandler;
+import com.farao_community.farao.data.crac_api.ExtensionsHandler;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import com.fasterxml.jackson.core.type.TypeReference;
