+++ data/crac/crac-result-extensions/pom.xml
@@ -27,11 +27,6 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>farao-crac-impl</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
             <artifactId>farao-crac-io-json</artifactId>
             <version>${project.version}</version>
         </dependency>
